+++ Helper/Config.php
@@ -1,37 +1,48 @@
 <?php
 /**
- * This file is part of CloudImage Responsive
+ * This file is part of Scaleflex Cloudimage
  *
  * @author Alyzeo LTD <info@alyzeo.com>
- * @category CloudImage
- * @package CloudImage\Responsive\Model
+ * @category Scaleflex
+ * @package Scaleflex\Cloudimage
  * @license BSD-3-Clause
  * @copyright Copyright (c) 2021 Cloudimage (https://www.cloudimage.io/)
  */
 
-namespace CloudImage\Responsive\Helper;
+namespace Scaleflex\Cloudimage\Helper;
 
 use Magento\Framework\App\Helper\AbstractHelper;
 use Magento\Store\Model\ScopeInterface;
 
+/**
+ * Class Config
+ *
+ * @package Scaleflex\Cloudimage\Helper
+ */
 class Config extends AbstractHelper
 {
-    const XML_PATH_CLOUDIMAGE_RESPONSIVE_GENERAL_ACTIVE = 'cloudimage_responsive/general/active';
-    const XML_PATH_CLOUDIMAGE_RESPONSIVE_GENERAL_TOKEN = 'cloudimage_responsive/general/token';
-    const XML_PATH_CLOUDIMAGE_RESPONSIVE_OPTIONS_LAZY_LOADING = 'cloudimage_responsive/options/lazy_loading';
-    const XML_PATH_CLOUDIMAGE_RESPONSIVE_OPTIONS_IGNORE_NODE_IMG_SIZE = 'cloudimage_responsive/options/ignore_node_img_size';
-    const XML_PATH_CLOUDIMAGE_RESPONSIVE_OPTIONS_IGNORE_STYLE_IMG_SIZE = 'cloudimage_responsive/options/ignore_style_img_size';
-    const XML_PATH_CLOUDIMAGE_RESPONSIVE_OPTIONS_DO_NOT_REPLACE_URL = 'cloudimage_responsive/options/do_not_replace_url';
-    const XML_PATH_CLOUDIMAGE_RESPONSIVE_CDN_API_URL = 'cloudimage_responsive/cdn/api_url';
-    const XML_PATH_CLOUDIMAGE_RESPONSIVE_ADVANCED_CUSTOM_FUNCTION_ACTIVE = 'cloudimage_responsive/advanced/custom_function_active';
-    const XML_PATH_CLOUDIMAGE_RESPONSIVE_ADVANCED_CUSTOM_FUNCTION = 'cloudimage_responsive/advanced/custom_function';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_GENERAL_ACTIVE = 'scaleflex_cloudimage/general/active';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_GENERAL_TOKEN = 'scaleflex_cloudimage/general/token';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_OPTIONS_LAZY_LOADING = 'scaleflex_cloudimage/options/lazy_loading';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_OPTIONS_IGNORE_NODE_IMG_SIZE = 'scaleflex_cloudimage/options/ignore_node_img_size';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_OPTIONS_IGNORE_STYLE_IMG_SIZE = 'scaleflex_cloudimage/options/ignore_style_img_size';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_OPTIONS_DO_NOT_REPLACE_URL = 'scaleflex_cloudimage/options/do_not_replace_url';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_CDN_API_URL = 'scaleflex_cloudimage/cdn/api_url';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_CUSTOM_FUNCTION_ACTIVE = 'scaleflex_cloudimage/advanced/custom_function_active';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_CUSTOM_FUNCTION = 'scaleflex_cloudimage/advanced/custom_function';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_IMAGE_QUALITY = 'scaleflex_cloudimage/advanced/image_quality';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_PROCESS_SVG = 'scaleflex_cloudimage/advanced/ignore_svg';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_LIBRARY_OPTIONS = 'scaleflex_cloudimage/advanced/library_options';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_REMOVE_V7 = 'scaleflex_cloudimage/advanced/remove_v7';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_FOTORAMA_COMPATIBILITY = 'scaleflex_cloudimage/advanced/fortorama_compatibility';
+    const XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_DEVICEPIXELRATIO = 'scaleflex_cloudimage/advanced/devicepixelratio';
 
     /**
      * @return bool
      */
     public function getApiUrl()
     {
-        return $this->scopeConfig->getValue(self::XML_PATH_CLOUDIMAGE_RESPONSIVE_CDN_API_URL);
+        return $this->scopeConfig->getValue(self::XML_PATH_SCALEFLEX_CLOUDIMAGE_CDN_API_URL);
     }
 
     /**
@@ -42,13 +53,13 @@ class Config extends AbstractHelper
     public function getToken()
     {
         return $this->scopeConfig->getValue(
-            self::XML_PATH_CLOUDIMAGE_RESPONSIVE_GENERAL_TOKEN,
+            self::XML_PATH_SCALEFLEX_CLOUDIMAGE_GENERAL_TOKEN,
             ScopeInterface::SCOPE_WEBSITE
         );
     }
 
     /**
-     * Get CloudImage Responsive plugin configuration
+     * Get Scaleflex Cloudimage plugin configuration
      *
      * @return array
      */
@@ -71,16 +82,47 @@ class Config extends AbstractHelper
     }
 
     /**
-     * Is custom function active
+     * Get custom function
      *
      * @return bool
      */
     public function getCustomFunction()
     {
         return $this->scopeConfig->getValue(
-            self::XML_PATH_CLOUDIMAGE_RESPONSIVE_ADVANCED_CUSTOM_FUNCTION,
+            self::XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_CUSTOM_FUNCTION,
             ScopeInterface::SCOPE_STORE
         );
+    }
+
+    /**
+     * Get Image Quality
+     *
+     * More info here: https://docs.cloudimage.io/go/cloudimage-documentation-v7/en/image-compression/image-formats
+     * @return int
+     */
+    public function getImageQuality()
+    {
+        $imageQuality = (int)$this->scopeConfig->getValue(
+            self::XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_IMAGE_QUALITY
+        );
+        if ($imageQuality > 0 && $imageQuality <= 100) {
+            return $imageQuality;
+        }
+        return 100;
+    }
+
+    /**
+     * Get Library Options
+     *
+     * @return string
+     */
+    public function getLibraryOptions()
+    {
+        $options = $this->scopeConfig->getValue(self::XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_LIBRARY_OPTIONS);
+        if (is_string($options) && strlen(trim($options)) > 0) {
+            return trim($options);
+        }
+        return '';
     }
 
     /**
@@ -91,7 +133,7 @@ class Config extends AbstractHelper
     public function isActive()
     {
         return (bool)$this->scopeConfig->getValue(
-            self::XML_PATH_CLOUDIMAGE_RESPONSIVE_GENERAL_ACTIVE,
+            self::XML_PATH_SCALEFLEX_CLOUDIMAGE_GENERAL_ACTIVE,
             ScopeInterface::SCOPE_STORE
         );
     }
@@ -103,10 +145,18 @@ class Config extends AbstractHelper
      */
     public function isCustomFunctionActive()
     {
-        return (bool)$this->scopeConfig->getValue(
-            self::XML_PATH_CLOUDIMAGE_RESPONSIVE_ADVANCED_CUSTOM_FUNCTION_ACTIVE,
+        $isCustomFunctionActive = (bool)$this->scopeConfig->getValue(
+            self::XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_CUSTOM_FUNCTION_ACTIVE,
             ScopeInterface::SCOPE_STORE
         );
+
+        if ($isCustomFunctionActive) {
+            $customFunction = $this->getCustomFunction();
+            if (is_string($customFunction) && strlen(trim($customFunction)) > 0) {
+                return true;
+            }
+        }
+        return false;
     }
 
     /**
@@ -115,7 +165,7 @@ class Config extends AbstractHelper
     public function isDoNotReplaceUrl()
     {
         return (bool)$this->scopeConfig->getValue(
-            self::XML_PATH_CLOUDIMAGE_RESPONSIVE_OPTIONS_DO_NOT_REPLACE_URL,
+            self::XML_PATH_SCALEFLEX_CLOUDIMAGE_OPTIONS_DO_NOT_REPLACE_URL,
             ScopeInterface::SCOPE_WEBSITE
         );
     }
@@ -126,7 +176,7 @@ class Config extends AbstractHelper
     public function isIgnoreNodeImgSize()
     {
         return (bool)$this->scopeConfig->getValue(
-            self::XML_PATH_CLOUDIMAGE_RESPONSIVE_OPTIONS_IGNORE_NODE_IMG_SIZE,
+            self::XML_PATH_SCALEFLEX_CLOUDIMAGE_OPTIONS_IGNORE_NODE_IMG_SIZE,
             ScopeInterface::SCOPE_WEBSITE
         );
     }
@@ -137,7 +187,7 @@ class Config extends AbstractHelper
     public function isIgnoreStyleImgSize()
     {
         return (bool)$this->scopeConfig->getValue(
-            self::XML_PATH_CLOUDIMAGE_RESPONSIVE_OPTIONS_IGNORE_STYLE_IMG_SIZE,
+            self::XML_PATH_SCALEFLEX_CLOUDIMAGE_OPTIONS_IGNORE_STYLE_IMG_SIZE,
             ScopeInterface::SCOPE_WEBSITE
         );
     }
@@ -150,8 +200,46 @@ class Config extends AbstractHelper
     public function isLazyLoading()
     {
         return (bool)$this->scopeConfig->getValue(
-            self::XML_PATH_CLOUDIMAGE_RESPONSIVE_OPTIONS_LAZY_LOADING,
+            self::XML_PATH_SCALEFLEX_CLOUDIMAGE_OPTIONS_LAZY_LOADING,
             ScopeInterface::SCOPE_WEBSITE
         );
+    }
+
+    /**
+     * Ignore SVG images ?
+     *
+     * @return bool
+     */
+    public function isIgnoreSvg()
+    {
+        return $this->scopeConfig->isSetFlag(self::XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_PROCESS_SVG);
+    }
+
+
+    /**
+     * Device pixel ratio
+     *
+     * @return string
+     */
+    public function getDevicePixelRatio()
+    {
+        return $this->scopeConfig->getValue(self::XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_DEVICEPIXELRATIO);
+    }
+
+    /**
+     * Ignore SVG images ?
+     *
+     * @return bool
+     */
+    public function isRemoveV7()
+    {
+        return $this->scopeConfig->isSetFlag(self::XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_REMOVE_V7);
+    }
+
+    /**
+     * @return bool
+     */
+    public function isFotoramaCompatibility() {
+        return $this->scopeConfig->isSetFlag(self::XML_PATH_SCALEFLEX_CLOUDIMAGE_ADVANCED_FOTORAMA_COMPATIBILITY);
     }
 }