+++ CHANGELOG.md
@@ -2,7 +2,33 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.2] - 2022-04-08
+
+* Add Fotorama Compatibility Mode option
+* Add Device Pixel Max Ratio option
+
+## [2.0.1] - 2022-02-01
+
+* Remove "/v7" from the URL part option added
+
+## [2.0.0] - 2022-01-17
+
+  * Add a text field in the configuration to modify the URL for retrieving the Cloudimage library
+  * Add an option in the back office to define the image compression (select box from 1 to 100)
+  * Add a Yes / No option to ignore SVG type images
+  * Increase the test before integration of the JS custom function
+  * Rename the Cloudimage_Responsive module to Scaleflex_Cloudimage
+  * Add a "lazy-off" class to avoid lazy loading on specified images
+
+## [1.1.0] - 2020-03-21
+
+  * Integration "on the fly" of Magento templates with replacement of image src attribute by ci-src
+  * Add CSP definition for cloudimg.io
+  * Tag script inline optimisation
+  * Fix typo CloudImage > Cloudimage
+  * Update documentation
+
 ## [1.0.0] - 2020-12-16
 
-  * First version of CloudImage Responsive plugin by Alyzeo LTD
+  * First version of Cloudimage Responsive plugin by Alyzeo LTD